refactor: 추천 관련 Mapper 분리
